One synopsis to rule them all
Benefits:
- Reduces number of files.
- Gives a single file from which one can learn the library.
