+++ README.md
@@ -1,3 +1,5 @@
 # [≡](#contents) [Lax.C++](#) [![Build Status](https://travis-ci.org/per-framework/lax.cpp.svg?branch=v1)](https://travis-ci.org/per-framework/lax.cpp)
 
 Non-strict template metaprogramming primitives for C++.
+
+See [`synopsis.hpp`](provides/include/lax_v1/synopsis.hpp) for the API.
