--- provides/include/lax_v1/arithmetic_synopsis.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class LhsExpr, class RhsExpr> struct add_m;
-template <class LhsExpr, class RhsExpr> struct div_m;
-template <class LhsExpr, class RhsExpr> struct mod_m;
-template <class LhsExpr, class RhsExpr> struct mul_m;
-template <class LhsExpr, class RhsExpr> struct sub_m;
-
-template <class Expr> struct neg_m;
-
-} // namespace lax_v1
