--- provides/include/lax_v1/comparison_synopsis.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class LhsExpr, class RhsExpr> struct gt_m;
-template <class LhsExpr, class RhsExpr> struct gte_m;
-template <class LhsExpr, class RhsExpr> struct lt_m;
-template <class LhsExpr, class RhsExpr> struct lte_m;
-
-} // namespace lax_v1
