+++ provides/include/lax_v1/force.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-#include "lax_v1/force_synopsis.hpp"
+#include "lax_v1/synopsis.hpp"
