--- provides/include/lax_v1/force_synopsis.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class Expr> using force_t = typename Expr::eval;
-
-}
