+++ provides/include/lax_v1/lazify.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "lax_v1/lazify_synopsis.hpp"
+#include "lax_v1/synopsis.hpp"
 #include "lax_v1/type.hpp"
 #include "lax_v1/value.hpp"
