--- provides/include/lax_v1/lazify_synopsis.hpp
@@ -1,12 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <template <class... Parameters> class Function, class... Arguments>
-struct lazify_t;
-
-template <class Type, template <class... Parameters> class Function,
-          class... Arguments>
-struct lazify_v;
-
-} // namespace lax_v1
