--- provides/include/lax_v1/logical_synopsis.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class... Exprs> struct if_m;
-
-template <class... Exprs> struct and_m;
-
-template <class Expr> struct not_m;
-
-template <class... Exprs> struct or_m;
-
-} // namespace lax_v1
