+++ provides/include/lax_v1/type.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "lax_v1/type_synopsis.hpp"
+#include "lax_v1/synopsis.hpp"
 
 template <class T> struct lax_v1::type_t {
   using eval = type_t;
