--- provides/include/lax_v1/type_synopsis.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class T> struct type_t;
-
-template <class Expr> using type_of_t = typename Expr::type;
-
-} // namespace lax_v1
