+++ provides/include/lax_v1/type_traits.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "lax_v1/lazify.hpp"
-#include "lax_v1/type_traits_synopsis.hpp"
+#include "lax_v1/synopsis.hpp"
 
 #include <cstddef>
 #include <type_traits>
