--- provides/include/lax_v1/type_traits_synopsis.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-namespace lax_v1 {
-
-template <class TypeExpr> struct alignment_of_m;
-
-template <class TypeExpr> struct is_array_m;
-
-template <class TypeExpr> struct is_pointer_m;
-
-template <class LhsTypeExpr, class RhsTypeExpr> struct is_same_m;
-
-template <class TypeExpr> struct remove_all_extents_m;
-
-template <class TypeExpr> struct remove_pointer_m;
-
-} // namespace lax_v1
