+++ provides/include/lax_v1/value.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "lax_v1/synopsis.hpp"
 #include "lax_v1/type.hpp"
-#include "lax_v1/value_synopsis.hpp"
 
 #include <type_traits>
